// src/app/models/emitente.ts
export class Emitente{
    
    id: number;
    nomeEmitente: string;
    cnpjEmitente: string;
    logradouroEmitente: string;
    numero_emitente: string;
    complementoEmitente: string;
    bairroEmitente: string;
    cepEmitente: string;
    telefoneEmitente: string;
    emailEmitente: string;
    logoEmitente: string;
    
}